task app done

+++ app/Http/Controllers/TasksController.php
@@ -14,16 +14,28 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
      
+     
+     
      // getでmessages/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all(); //Taskはmodel名。all()はすべてのデータを取得。
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [ //tasks.indexはbladeファイル名
-            'tasks' => $tasks, //$tasksには取得したall()のデータが入っている。これを'tasks'に代入。'tasks'はbladeファイル内で$tasksとして使用できる。
-        ]);
+        if (\Auth::check()) { // 認証済みの場合
+             // 認証済みユーザを取得
+            $user = \Auth::user();
+            
+            // task一覧を取得
+            $tasks = $user->tasks()->get(); //ユーザのtaskのみ取得
+            
+            // task一覧をviewで表示
+            return view('tasks.index', [ //tasks.indexはbladeファイル名
+                'tasks' => $tasks, //$tasksには取得したtaskデータが入っている。これを'tasks'に代入。'tasks'はbladeファイル内で$tasksとして使用できる。
+            ]);
+            
+        }
+         
+         //loginしていなかったらwelcome.blade.phpにリダイレクト
+         return view('welcome');
+         
     }
 
     /**
@@ -39,6 +51,7 @@ class TasksController extends Controller
         return view('tasks.create', [
             'task' => $task,
         ]);
+        
     }
 
     /**
@@ -55,11 +68,11 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）のtaskとして作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -73,10 +86,10 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値でtaskを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+        // task詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -135,8 +148,11 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、taskを削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
